+++ src/joseph3d_back_tof_lm_cuda.cu
@@ -41,7 +41,7 @@ __global__ void joseph3d_back_tof_lm_cuda_kernel(float *xstart,
                                                  float tofbin_width,
                                                  float *sigma_tof,
                                                  float *tofcenter_offset,
-                                                 int n_sigmas,
+                                                 float n_sigmas,
                                                  int *tof_bin)
 {
   long long i = blockDim.x * blockIdx.x + threadIdx.x;
@@ -505,7 +505,7 @@ extern "C" void joseph3d_back_tof_lm_cuda(float *h_xstart,
                                             float tofbin_width,
                                             float *h_sigma_tof,
                                             float *h_tofcenter_offset,
-                                            int n_sigmas,
+                                            float n_sigmas,
                                             int *h_tof_bin,
                                             int threadsperblock,
                                             int num_devices)
